Funcion nueva de mostrar fixture

+++ src/main/java/com/patricio/springboot/app/repository/ZonaRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface ZonaRepository extends JpaRepository<Zona, Integer> {
+public interface ZonaRepository extends JpaRepository<Zona, Long> {
     Optional<Zona> findById(Long id);
     List<Zona> findByTorneoId(Long torneoId);
